requirements.txt
apscheduler
arrow
attrs
beautifulsoup4
selenium
strictyaml
tabulate
xdg
